+++ src/features/schedule/components/schedule-form/schedule-form.tsx
@@ -1,0 +1,81 @@
+import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from "@/components/ui/form";
+import { DAYS } from "@/constants/days";
+import FieldAvailabilityDay from "./field-availability-day";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { SheetFooter, SheetClose } from "@/components/ui/sheet";
+import { useScheduleForm } from "../../hooks/use-schedule-form";
+
+const ScheduleForm = () => {
+  const { form, fieldArray, addFieldAvailability, onSubmit } = useScheduleForm();
+
+  const {
+    control,
+    handleSubmit,
+    formState: { isSubmitting, isValidating },
+  } = form;
+
+  const { fields, remove } = fieldArray;
+
+  return (
+    <Form {...form}>
+      <form onSubmit={handleSubmit(onSubmit)}>
+        <div className="space-y-6 px-4 mb-4">
+          {DAYS.map((day) => {
+            const dayFields = fields
+              .map((field, idx) => ({ ...field, originalIndex: idx }))
+              .filter((field) => field.day === day);
+            return (
+              <FieldAvailabilityDay
+                key={day}
+                day={day}
+                dayFields={dayFields}
+                addFieldAvailability={addFieldAvailability}
+                remove={remove}
+                control={control}
+              />
+            );
+          })}
+          <FormField
+            control={control}
+            name="duration"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Duration</FormLabel>
+                <FormControl>
+                  <Input type="number" min={15} max={60} step={15} disabled {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+          <FormField
+            control={control}
+            name="maximumPlayerCount"
+            render={({ field }) => (
+              <FormItem>
+                <FormLabel>Maximum Player Count</FormLabel>
+                <FormControl>
+                  <Input type="number" disabled {...field} />
+                </FormControl>
+                <FormMessage />
+              </FormItem>
+            )}
+          />
+        </div>
+        <SheetFooter>
+          <Button type="submit" disabled={isSubmitting || isValidating}>
+            Schedule Players
+          </Button>
+          <SheetClose asChild>
+            <Button type="button" variant="outline" disabled={isSubmitting || isValidating}>
+              Close
+            </Button>
+          </SheetClose>
+        </SheetFooter>
+      </form>
+    </Form>
+  );
+};
+
+export default ScheduleForm;
