+++ src/features/schedule/hooks/use-training-blocks.ts
