+++ src/routes/schedule.tsx
@@ -1,7 +1,7 @@
 import Calendar from "@/features/schedule/components/calendar/calendar";
-import ScheduleSheet from "@/features/schedule/components/field-availability-form/schedule-sheet";
+import ScheduleSheet from "@/features/schedule/components/schedule-form/schedule-sheet";
 import ScheduleActionBar from "@/features/schedule/components/schedule-action-bar";
-import { useScheduleSheet } from "@/features/schedule/hooks/useScheduleSheet";
+import { useScheduleSheet } from "@/features/schedule/hooks/use-schedule-sheet";
 import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/schedule")({
@@ -9,12 +9,12 @@ export const Route = createFileRoute("/schedule")({
 });
 
 function Schedule() {
-  const scheduleSheetReturn = useScheduleSheet([]);
-  const { unassignedPlayerNames, openScheduleSheet } = scheduleSheetReturn;
+  const scheduleSheetReturn = useScheduleSheet();
+  const { openScheduleSheet } = scheduleSheetReturn;
 
   return (
     <>
-      <ScheduleActionBar openScheduleSheet={openScheduleSheet} unassignedPlayerNames={unassignedPlayerNames} />
+      <ScheduleActionBar openScheduleSheet={openScheduleSheet} unassignedPlayerNames={[]} />
       <Calendar players={[]} trainingBlocks={[]} />
       <ScheduleSheet {...scheduleSheetReturn} />
     </>
